fixed messages so that new messages are bold
