+++ app/scss/_Message.scss
@@ -5,3 +5,7 @@
 .messageContainer {
     font-weight: 400;
 }
+
+.avatar {
+    object-fit: cover;
+}
